# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Gymbro API", "version": "0.1.0"}, "paths": {"/users/": {"post": {"summary": "Create User", "operationId": "create_user_users__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"summary": "Read Users", "operationId": "read_users_users__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/User"}, "title": "Response Read Users Users  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}": {"delete": {"summary": "Delete User", "operationId": "delete_user_users__user_id__delete", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/routines/": {"post": {"summary": "Create Routine", "operationId": "create_routine_routines__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoutineCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Routine"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/routines/student/{student_id}": {"get": {"summary": "Read Routines For Student", "operationId": "read_routines_for_student_routines_student__student_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "student_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Student Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Routine"}, "title": "Response Read Routines For Student Routines Student  Student Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/payments/": {"post": {"summary": "Create Payment", "operationId": "create_payment_payments__post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaymentCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Payment"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"summary": "Read Payments", "operationId": "read_payments_payments__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Payment"}, "title": "Response Read Payments Payments  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/login": {"post": {"summary": "Login", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLogin"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Exercise": {"properties": {"name": {"type": "string", "title": "Name"}, "sets": {"type": "integer", "title": "Sets", "default": 3}, "reps": {"type": "integer", "title": "Reps", "default": 10}, "weight": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Weight"}, "id": {"type": "integer", "title": "Id"}, "routine_id": {"type": "integer", "title": "Routine Id"}}, "type": "object", "required": ["name", "id", "routine_id"], "title": "Exercise"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Payment": {"properties": {"amount": {"type": "number", "title": "Amount"}, "payment_method": {"type": "string", "title": "Payment Method"}, "id": {"type": "integer", "title": "Id"}, "student_id": {"type": "integer", "title": "Student Id"}, "date": {"type": "string", "format": "date-time", "title": "Date"}}, "type": "object", "required": ["amount", "payment_method", "id", "student_id", "date"], "title": "Payment"}, "PaymentCreate": {"properties": {"amount": {"type": "number", "title": "Amount"}, "payment_method": {"type": "string", "title": "Payment Method"}, "student_id": {"type": "integer", "title": "Student Id"}}, "type": "object", "required": ["amount", "payment_method", "student_id"], "title": "PaymentCreate"}, "Routine": {"properties": {"day_of_week": {"type": "integer", "title": "Day Of Week"}, "name": {"type": "string", "title": "Name"}, "id": {"type": "integer", "title": "Id"}, "student_id": {"type": "integer", "title": "Student Id"}, "coach_id": {"type": "integer", "title": "Coach Id"}, "exercises": {"items": {"$ref": "#/components/schemas/Exercise"}, "type": "array", "title": "Exercises", "default": []}}, "type": "object", "required": ["day_of_week", "name", "id", "student_id", "coach_id"], "title": "Routine"}, "RoutineCreate": {"properties": {"day_of_week": {"type": "integer", "title": "Day Of Week"}, "name": {"type": "string", "title": "Name"}, "student_id": {"type": "integer", "title": "Student Id"}, "coach_id": {"type": "integer", "title": "Coach Id"}}, "type": "object", "required": ["day_of_week", "name", "student_id", "coach_id"], "title": "RoutineCreate"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "User": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "is_coach": {"type": "boolean", "title": "Is Coach", "default": false}, "id": {"type": "integer", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["email", "name", "id", "created_at"], "title": "User"}, "UserCreate": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "is_coach": {"type": "boolean", "title": "Is Coach", "default": false}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "name", "password"], "title": "UserCreate"}, "UserLogin": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserLogin"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}